requirements.txt
slither-analyzer>=0.9.0
graphviz>=0.20.0
click>=8.0.0
matplotlib>=3.5.0
networkx>=2.8.0
pytest>=7.0.0